+++ teams-js-app-sdk/test/public/config.spec.ts
@@ -1,6 +1,6 @@
 import { pages } from '../../src/public/pages';
 import { Utils } from '../utils';
-import { core } from '../../src/public/publicAPIs';
+import { app } from '../../src/public/app';
 
 describe('config', () => {
   // Use to send a mock message from the app.
@@ -16,8 +16,8 @@ describe('config', () => {
 
   afterEach(() => {
     // Reset the object since it's a singleton
-    if (core._uninitialize) {
-      core._uninitialize();
+    if (app._uninitialize) {
+      app._uninitialize();
     }
   });