+++ teams-js-app-sdk/test/public/dialog.spec.ts
@@ -2,7 +2,7 @@ import { DialogInfo } from '../../src/public/interfaces';
 import { DialogDimension } from '../../src/public/constants';
 import { dialog } from '../../src/public/dialog';
 import { Utils } from '../utils';
-import { core } from '../../src/public/publicAPIs';
+import { app } from '../../src/public/app';
 
 describe('Dialog', () => {
   // Use to send a mock message from the app.
@@ -18,8 +18,8 @@ describe('Dialog', () => {
 
   afterEach(() => {
     // Reset the object since it's a singleton
-    if (core._uninitialize) {
-      core._uninitialize();
+    if (app._uninitialize) {
+      app._uninitialize();
     }
   });