+++ teams-js-app-sdk/test/public/people.spec.ts
@@ -1,5 +1,5 @@
 import { FramelessPostMocks } from '../framelessPostMocks';
-import { core } from '../../src/public/publicAPIs';
+import { app } from '../../src/public/app';
 import { FrameContexts } from '../../src/public/constants';
 import { DOMMessageEvent } from '../../src/internal/interfaces';
 import { SdkError, ErrorCode } from '../../src/public/interfaces';
@@ -12,18 +12,18 @@ describe('peoplePicker', () => {
   const mobilePlatformMock = new FramelessPostMocks();
   const minVersionForSelectPeople = '2.0.0';
   const originalDefaultPlatformVersion = '1.6.0';
- 
+
   beforeEach(() => {
     mobilePlatformMock.messages = [];
 
     // Set a mock window for testing
-    core._initialize(mobilePlatformMock.mockWindow);
+    app._initialize(mobilePlatformMock.mockWindow);
   });
 
   afterEach(() => {
     // Reset the object since it's a singleton
-    if (core._uninitialize) {
-      core._uninitialize();
+    if (app._uninitialize) {
+      app._uninitialize();
     }
   });
 
@@ -33,9 +33,7 @@ describe('peoplePicker', () => {
    * People Picker tests
    */
   it('should not allow selectPeople calls with null callback', () => {
-    expect(() => people.selectPeople(null)).toThrowError(
-      '[people picker] Callback cannot be null',
-    );
+    expect(() => people.selectPeople(null)).toThrowError('[people picker] Callback cannot be null');
   });
 
   it('should allow selectPeople calls with null peoplePickerInputs', async () => {
@@ -111,17 +109,19 @@ describe('peoplePicker', () => {
     expect(message.args.length).toBe(1);
 
     const callbackId = message.id;
-    const result = [{
-     objectId: "5842943a-aa5a-470a-bfdc-7311b9988962",
-     displayName: "Sonal Jha",
-     email: "sojh@m365x347208.onmicrosoft.com"
-    } as people.PeoplePickerResult];
+    const result = [
+      {
+        objectId: '5842943a-aa5a-470a-bfdc-7311b9988962',
+        displayName: 'Sonal Jha',
+        email: 'sojh@m365x347208.onmicrosoft.com',
+      } as people.PeoplePickerResult,
+    ];
     mobilePlatformMock.respondToMessage({
       data: {
         id: callbackId,
-        args: [undefined, result]
-      }
-    } as DOMMessageEvent)
+        args: [undefined, result],
+      },
+    } as DOMMessageEvent);
 
     expect(peoplePickerError).toBeFalsy();
     expect(result.length).toBe(1);
@@ -141,12 +141,12 @@ describe('peoplePicker', () => {
     mobilePlatformMock.setClientSupportedSDKVersion(minVersionForSelectPeople);
     let peoplePickerResult: people.PeoplePickerResult[], peoplePickerError: SdkError;
     const peoplePickerInput: people.PeoplePickerInputs = {
-     title: "Hello World",
-     setSelected: null,
-     openOrgWideSearchInChatOrChannel: true,
-     singleSelect: true
+      title: 'Hello World',
+      setSelected: null,
+      openOrgWideSearchInChatOrChannel: true,
+      singleSelect: true,
     };
-    people.selectPeople( (e: SdkError, m: people.PeoplePickerResult[]) => {
+    people.selectPeople((e: SdkError, m: people.PeoplePickerResult[]) => {
       peoplePickerError = e;
       peoplePickerResult = m;
     }, peoplePickerInput);
@@ -159,9 +159,9 @@ describe('peoplePicker', () => {
     mobilePlatformMock.respondToMessage({
       data: {
         id: callbackId,
-        args: [{ errorCode: ErrorCode.INTERNAL_ERROR }]
-      }
-    } as DOMMessageEvent)
+        args: [{ errorCode: ErrorCode.INTERNAL_ERROR }],
+      },
+    } as DOMMessageEvent);
 
     expect(peoplePickerResult).toBeFalsy();
     expect(peoplePickerError.errorCode).toBe(ErrorCode.INTERNAL_ERROR);