+++ teams-js-app-sdk/test/public/teamsAPIs.spec.ts
@@ -1,5 +1,5 @@
 import { TabInstanceParameters, FrameInfo } from '../../src/public/interfaces';
-import { core } from '../../src/public/publicAPIs';
+import { app, core } from '../../src/public/app';
 import { teamsCore } from '../../src/public/teamsAPIs';
 import { pages } from '../../src/public/pages';
 import { Utils } from '../utils';
@@ -16,13 +16,13 @@ describe('teamsjsAppSDK-TeamsAPIs', () => {
     utils.childWindow.closed = false;
 
     // Set a mock window for testing
-    core._initialize(utils.mockWindow);
+    app._initialize(utils.mockWindow);
   });
 
   afterEach(() => {
     // Reset the object since it's a singleton
-    if (core._uninitialize) {
-      core._uninitialize();
+    if (app._uninitialize) {
+      app._uninitialize();
     }
   });
 
@@ -130,7 +130,7 @@ describe('teamsjsAppSDK-TeamsAPIs', () => {
 
   it("Ctrl+P shouldn't call print handler if printCapabilty is disabled", async () => {
     let handlerCalled = false;
-    core.initialize();
+    app.initialize();
     spyOn(teamsCore, 'print').and.callFake((): void => {
       handlerCalled = true;
     });
@@ -146,7 +146,7 @@ describe('teamsjsAppSDK-TeamsAPIs', () => {
 
   it("Cmd+P shouldn't call print handler if printCapabilty is disabled", async () => {
     let handlerCalled = false;
-    core.initialize();
+    app.initialize();
     spyOn(teamsCore, 'print').and.callFake((): void => {
       handlerCalled = true;
     });
@@ -162,7 +162,7 @@ describe('teamsjsAppSDK-TeamsAPIs', () => {
 
   it('print handler should successfully call default print handler', async () => {
     let handlerCalled = false;
-    core.initialize();
+    app.initialize();
     teamsCore.enablePrintCapability();
     spyOn(window, 'print').and.callFake((): void => {
       handlerCalled = true;
@@ -175,7 +175,7 @@ describe('teamsjsAppSDK-TeamsAPIs', () => {
 
   it('Ctrl+P should successfully call print handler', async () => {
     let handlerCalled = false;
-    core.initialize();
+    app.initialize();
     teamsCore.enablePrintCapability();
     spyOn(window, 'print').and.callFake((): void => {
       handlerCalled = true;
@@ -192,7 +192,7 @@ describe('teamsjsAppSDK-TeamsAPIs', () => {
 
   it('Cmd+P should successfully call print handler', async () => {
     let handlerCalled = false;
-    core.initialize();
+    app.initialize();
     teamsCore.enablePrintCapability();
     spyOn(window, 'print').and.callFake((): void => {
       handlerCalled = true;