+++ teams-js-app-sdk/test/utils.ts
@@ -1,4 +1,4 @@
-import { core } from '../src/public/publicAPIs';
+import { app } from '../src/public/app';
 import { GlobalVars } from '../src/internal/globalVars';
 import { defaultSDKVersionForCompatCheck } from '../src/internal/constants';
 import { DOMMessageEvent, ExtendedWindow } from '../src/internal/interfaces';
@@ -104,8 +104,8 @@ export class Utils {
     hostClientType?: string,
     validMessageOrigins?: string[],
   ): Promise<void> => {
-    core._initialize(this.mockWindow);
-    const promise = core.initialize(validMessageOrigins);
+    app._initialize(this.mockWindow);
+    const promise = app.initialize(validMessageOrigins);
 
     const initMessage = this.findMessageByFunc('initialize');
     expect(initMessage).not.toBeNull();
@@ -117,7 +117,7 @@ export class Utils {
 
   public initializeAsFrameless = (validMessageOrigins?: string[]): Promise<void> => {
     this.mockWindow.parent = null;
-    return core.initialize(validMessageOrigins);
+    return app.initialize(validMessageOrigins);
   };
 
   public findMessageByFunc = (func: string): MessageRequest => {