+++ utilities/teams-js-perf-test-app/src/App.tsx
@@ -1,11 +1,11 @@
 import React, { ReactElement } from 'react';
 import './App.css';
-import { core, appInitialization } from '@microsoft/teamsjs-app-sdk';
+import { app } from '@microsoft/teamsjs-app-sdk';
 import AppInitialization from './components/AppInitialization';
 
-core.initialize();
-appInitialization.notifyAppLoaded();
-appInitialization.notifySuccess();
+app.initialize();
+app.notifyAppLoaded();
+app.notifySuccess();
 
 export const noHubSdkMsg = ' was called, but there was no response from the Hub SDK.';